create sphinx-gallery rectilinear section
