+++ src/geovista/examples/rectilinear/README.rst
@@ -1,0 +1,2 @@
+Rectilinear Grids
+=================
