+++ src/geovista/registry.txt
@@ -37,8 +37,6 @@ tests/images/curvilinear__from_2d__orca_moll.png 576dd061b242aa2fd55166381aed91f
 tests/images/curvilinear__from_2d__orca.png 8552e2010bb058f190694cc7da3d096988edb30406b2d9ad1d64165eaa2f836a
 tests/images/earthquakes.png 2995a9277dfbc0505a6e175a5affa418a066ad39482c3de98678fe8521d56aa6
 tests/images/earthquakes_wink1.png bb929dc99d2ea390a217929a9217fc7b1e7828639cd9a0fcc51484b342b88d6f
-tests/images/from_1d__oisst_eqc.png 5e8c62933238d08baaa09eb6f12f7d5e47e731023bbe8600b75dc96c1ef21971
-tests/images/from_1d__oisst.png 92da94eb490eaa72acca8938e22edacf530b43a15bd215633bb6111d5380bfba
 tests/images/from_1d__synthetic_face_m1_n1.png 917dca4452ab0c2fbe1fccfe8439855f3cd1becb8b73c2424086a622d3867c62
 tests/images/from_1d__synthetic_face_m1_n1_robin.png 915c3122343eafe1de65e0c02980595886f2afa3dfd1b1610a98b218de358186
 tests/images/from_1d__synthetic_node_m1_n1_moll.png e6bb80546c4131f6431736746967b155240e0ef741767aa2a8dbd88adbd0bfaa
@@ -66,6 +64,8 @@ tests/images/from_unstructured__tri_hammer.png 722c9e22881254547d415c2acc9a9b2fd
 tests/images/from_unstructured__tri.png 351847600795a0c1569e71414786311091d1f520915452c3def67b7b03a9244b
 tests/images/point_cloud__from_points__orca_cloud_eqc.png 6421665bc4bff1128b2876b7b83978b3c6e894be607b39764e136f013b795ddc
 tests/images/point_cloud__from_points__orca_cloud.png a5fa2a43eb6f8b1ab2adf6f6e757830b07305a87a3bb9abf64c48d2d2d8a066b
+tests/images/rectilinear__from_1d__oisst.png 92da94eb490eaa72acca8938e22edacf530b43a15bd215633bb6111d5380bfba
+tests/images/rectilinear__from_1d__oisst_eqc.png 5e8c62933238d08baaa09eb6f12f7d5e47e731023bbe8600b75dc96c1ef21971
 tests/images/spatial_index__uber_h3.png 08cb9cbb5f3ae41d1955efe7cdcfac999b2c6e67072f57871dde85ab49205289
 tests/images/vector_data__vectors.png fdfea84491039fa742b73e610b834272d73eac7a70ab6b8c6e9b51019c0e1a95
 tos_Omon_AWI-ESM-1-1-LR_historical_r1i1p1f1_gn_185001-185012.nc 28a8f7b8f3d8edded39a7b9c93afe54aeefbf271086927c7f684207327acc882
